# [14.0][FIX] fix performance issue with dummy fiscal records