--- a/l10n_br_account/models/account_invoice_line.py
+++ b/l10n_br_account/models/account_invoice_line.py
@@ -174,14 +174,7 @@ class AccountMoveLine(models.Model):
                 )
             )
 
-        recompute_line_after_id = self.search([], order="id DESC", limit=1).id
-        lines = super(
-            AccountMoveLine,
-            self.with_context(
-                recompute_account_move_line_after_id=recompute_line_after_id
-            ),
-        ).create(vals_list)
-
+        lines = super().create(vals_list)
         for line in lines.filtered(lambda l: l.fiscal_document_line_id != dummy_line):
             shadowed_fiscal_vals = line._prepare_shadowed_fields_dict()
             doc_id = line.move_id.fiscal_document_id.id