fix para https://github.com/OCA/l10n-brazil/pull/2414

talvez pode ser simplificado ainda...

```
2023-04-05 06:31:25,559 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 0 of 1000. time: 0.373429 seconds 
2023-04-05 06:31:25,872 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 1 of 1000. time: 0.312446 seconds 
2023-04-05 06:31:26,175 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 2 of 1000. time: 0.303476 seconds 
2023-04-05 06:31:26,456 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 3 of 1000. time: 0.280340 seconds 
2023-04-05 06:31:26,759 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 4 of 1000. time: 0.303494 seconds 
2023-04-05 06:31:27,064 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 5 of 1000. time: 0.304760 seconds 
2023-04-05 06:31:27,378 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 6 of 1000. time: 0.313791 seconds 
2023-04-05 06:31:27,681 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 7 of 1000. time: 0.302083 seconds 
2023-04-05 06:31:27,984 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 8 of 1000. time: 0.303232 seconds 
2023-04-05 06:31:28,292 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 9 of 1000. time: 0.307639 seconds 
2023-04-05 06:31:28,590 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 10 of 1000. time: 0.297642 seconds 
2023-04-05 06:31:28,889 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 11 of 1000. time: 0.299162 seconds 
2023-04-05 06:31:29,201 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 12 of 1000. time: 0.312339 seconds 
2023-04-05 06:31:29,510 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 13 of 1000. time: 0.308371 seconds 
2023-04-05 06:31:29,800 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 14 of 1000. time: 0.289663 seconds 
2023-04-05 06:31:30,113 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 15 of 1000. time: 0.313165 seconds 
2023-04-05 06:31:30,427 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 16 of 1000. time: 0.314071 seconds 
2023-04-05 06:31:30,734 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 17 of 1000. time: 0.306985 seconds 
2023-04-05 06:31:31,025 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 18 of 1000. time: 0.290414 seconds 
2023-04-05 06:31:31,333 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 19 of 1000. time: 0.308137 seconds 
2023-04-05 06:31:31,623 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 20 of 1000. time: 0.289549 seconds 

[...]

2023-04-05 06:41:00,324 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 970 of 1000. time: 0.879219 seconds 
2023-04-05 06:41:01,277 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 971 of 1000. time: 0.952055 seconds 
2023-04-05 06:41:02,246 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 972 of 1000. time: 0.969211 seconds 
2023-04-05 06:41:03,130 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 973 of 1000. time: 0.883403 seconds 
2023-04-05 06:41:04,093 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 974 of 1000. time: 0.963713 seconds 
2023-04-05 06:41:05,010 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 975 of 1000. time: 0.916848 seconds 
2023-04-05 06:41:05,970 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 976 of 1000. time: 0.959778 seconds 
2023-04-05 06:41:06,809 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 977 of 1000. time: 0.838578 seconds 
2023-04-05 06:41:07,705 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 978 of 1000. time: 0.896061 seconds 
2023-04-05 06:41:08,665 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 979 of 1000. time: 0.959732 seconds 
2023-04-05 06:41:09,625 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 980 of 1000. time: 0.959845 seconds 
2023-04-05 06:41:10,622 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 981 of 1000. time: 0.996126 seconds 
2023-04-05 06:41:11,583 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 982 of 1000. time: 0.960982 seconds 
2023-04-05 06:41:12,445 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 983 of 1000. time: 0.861705 seconds 
2023-04-05 06:41:13,352 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 984 of 1000. time: 0.907475 seconds 
2023-04-05 06:41:14,359 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 985 of 1000. time: 1.006355 seconds 
2023-04-05 06:41:15,271 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 986 of 1000. time: 0.912321 seconds 
2023-04-05 06:41:16,241 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 987 of 1000. time: 0.969248 seconds 
2023-04-05 06:41:17,221 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 988 of 1000. time: 0.979583 seconds 
2023-04-05 06:41:18,164 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 989 of 1000. time: 0.943569 seconds 
2023-04-05 06:41:18,997 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 990 of 1000. time: 0.832385 seconds 
2023-04-05 06:41:19,962 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 991 of 1000. time: 0.965147 seconds 
2023-04-05 06:41:20,864 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 992 of 1000. time: 0.901376 seconds 
2023-04-05 06:41:21,841 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 993 of 1000. time: 0.977083 seconds 
2023-04-05 06:41:22,780 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 994 of 1000. time: 0.938937 seconds 
2023-04-05 06:41:23,719 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 995 of 1000. time: 0.938317 seconds 
2023-04-05 06:41:24,681 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 996 of 1000. time: 0.962555 seconds 
2023-04-05 06:41:25,728 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 997 of 1000. time: 1.046403 seconds 
2023-04-05 06:41:26,722 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 998 of 1000. time: 0.993669 seconds 
2023-04-05 06:41:27,697 300 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 999 of 1000. time: 0.975223 seconds
```

@antoniospneto @felipemotter a performance ja vai se degradando com a metade do que era antes. Porem eu ainda suspeito que isso é porque a gente ta criando tudo num loop so sem commit, algo que não é muito representativo de um uso real. Eu vou fazer um outro PR com ume commit no loop para ver melhor...