@rvalyi 2023-04-09
eu dei um revert no primeiro fix porque o segundo ja foi suficiente para matar completamente o problema, vcs podem verificar nos logs que os tempos ficaram igual ao que eu falei no primeiro comentário.

Eu expliquei um pouco mais no PR do fix final o porque o tempo ainda aumenta um pouco quando não se da commit no meio da criação dos registros https://github.com/OCA/l10n-brazil/pull/2422